+++ backend/src/domain/results/next_session_results.rs
